NLopt library for optimization step (pricing policy).
To allow for greater flexibility in the choice of functional form for
the hazard, an NLopt optimization step is given as an option
(derivative-based in this version), apart from the GSL root-finding
routine. Also, I include separate functions that define the hazards, so
it’s easier to change functional forms and not update every step where
they are used.